feat: split tests into different files
