+++ tests_zemu/tests/dkg.test.ts
@@ -19,13 +19,11 @@ import { defaultOptions, identities, models, restoreKeysTestCases } from './comm
 import IronfishApp, { IronfishKeys } from '@zondax/ledger-ironfish'
 import { isValidPublicAddress, multisig, UnsignedTransaction, verifyTransactions } from '@ironfish/rust-nodejs'
 import { Transaction } from '@ironfish/sdk'
-import { buildTx, minimizeRound3Inputs, startTextFn } from './utils'
+import { buildTx, minimizeRound3Inputs, runMethod, startTextFn } from './utils'
 import { TModel } from '@zondax/zemu/dist/types'
 import aggregateSignatureShares = multisig.aggregateSignatureShares
 
 jest.setTimeout(4500000)
-
-// Not sure about the start text for flex and stax, so we go with what it always work, which is the app name.
 
 // ONE_GLOBAL_APP: Use this flag if the whole DKG process will run in only one app (all participants, all rounds). This takes precedence over ONE_APP_PER_PARTICIPANT.
 // ONE_APP_PER_PARTICIPANT: Use this flag if the whole DKG process will run in one app per participant
@@ -36,47 +34,6 @@ const ONE_APP_PER_PARTICIPANT = 1
 const SKIP_ERRORS_IN_PHASE = true
 
 // Reference taken from https://github.com/iron-fish/ironfish/pull/5324/files
-
-const checkSimRequired = (m: IDeviceModel, sims: Zemu[], i: number): { sim: Zemu; created: boolean } => {
-  let created = false
-  let sim: Zemu | undefined
-
-  if (!sims.length) {
-    sim = new Zemu(m.path)
-    created = true
-  } else if (sims.length === 1) {
-    sim = sims[0]
-  } else {
-    sim = sims[i]
-  }
-
-  if (!sim) throw new Error('sim should have a value here')
-  return { sim, created }
-}
-
-const runMethod = async (m: IDeviceModel, rcvSims: Zemu[], i: number, fn: (sim: Zemu, app: IronfishApp) => Promise<any>): Promise<any> => {
-  const { sim, created } = checkSimRequired(m, rcvSims, i)
-
-  try {
-    if (created)
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-    const app = new IronfishApp(sim.getTransport(), true)
-    const resp = await fn(sim, app)
-
-    // Clean events from previous commands as each sim lives for many commands (DKG generation + signing)
-    await sim.deleteEvents()
-
-    return resp
-  } finally {
-    if (created) await sim.close()
-  }
-}
 
 describe.each(models)('DKG', function (m) {
   it.concurrent(`${m.name} - can start and stop container`, async function () {
@@ -421,524 +378,4 @@ describe.each(models)('DKG', function (m) {
       }
     })
   })
-
-  describe.each(restoreKeysTestCases)(
-    `${m.name} - restore keys`,
-    ({ index, encrypted, publicAddress, identities, proofKeys, viewKeys, publicPackage }) => {
-      test.concurrent(index + '', async () => {
-        for (let i = 0; i < encrypted.length; i++) {
-          const e = encrypted[i]
-
-          const sim = new Zemu(m.path)
-          try {
-            await sim.start({
-              ...defaultOptions,
-              model: m.name,
-              startText: startTextFn(m.name),
-              approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-              approveAction: ButtonKind.ApproveTapButton,
-            })
-            const app = new IronfishApp(sim.getTransport(), true)
-
-            // Restore keys
-            let respReq: any = app.dkgRestoreKeys(e)
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-${index}-restore-keys`)
-            let resp = await respReq
-            await sim.deleteEvents()
-
-            // Backup restored keys to compare snapshots for this process as it is deterministic (fixed keys)
-            respReq = app.dkgBackupKeys()
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-${index}-backup-keys`)
-            resp = await respReq
-            await sim.deleteEvents()
-
-            // Generate keys from the restored package to check we are generating the same keys when they were generated
-            respReq = app.dkgRetrieveKeys(IronfishKeys.ViewKey, true)
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-${index}-view-keys`)
-            resp = await respReq
-            await sim.deleteEvents()
-
-            expect(resp.viewKey.toString('hex')).toEqual(viewKeys.viewKey)
-            expect(resp.ovk.toString('hex')).toEqual(viewKeys.ovk)
-            expect(resp.ivk.toString('hex')).toEqual(viewKeys.ivk)
-
-            respReq = app.dkgRetrieveKeys(IronfishKeys.ProofGenerationKey, true)
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-${index}-proof-keys`)
-            resp = await respReq
-            await sim.deleteEvents()
-
-            expect(resp.ak.toString('hex')).toEqual(proofKeys.ak)
-            expect(resp.nsk.toString('hex')).toEqual(proofKeys.nsk)
-
-            respReq = app.dkgRetrieveKeys(IronfishKeys.PublicAddress, true)
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-${index}-public-addr`)
-            resp = await respReq
-            await sim.deleteEvents()
-
-            expect(resp.publicAddress.toString('hex')).toEqual(publicAddress)
-
-            respReq = app.dkgRetrieveKeys(IronfishKeys.DkgIdentity, true)
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            try {
-              await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-${index}-identity`)
-            } catch (e) {
-              // TODO navigate and approve, but do not compare snapshots... needs to be added to zemu
-              // Skip error, as a new public address is generated each time. Snapshots will be different in every run
-            }
-            resp = await respReq
-            await sim.deleteEvents()
-
-            expect(resp.identity.toString('hex')).toEqual(identities[i])
-
-            resp = await app.dkgGetPublicPackage()
-
-            expect(resp.publicPackage.toString('hex')).toEqual(publicPackage)
-
-            resp = await app.dkgGetIdentities()
-
-            const identitiesStr = resp.identities.map((i: any) => i.toString('hex'))
-            identities.forEach(i => expect(identitiesStr.includes(i)).toBeTruthy())
-          } finally {
-            await sim.close()
-          }
-        }
-      })
-    },
-  )
-
-  describe.each(restoreKeysTestCases)(`${m.name} - sign transaction`, ({ index, encrypted }) => {
-    test(index + '', async () => {
-      const participants = encrypted.length
-      const globalSims: Zemu[] = []
-
-      let identities: any[] = []
-      let commitments: any[] = []
-      let signatures: any[] = []
-
-      if (ONE_GLOBAL_APP) globalSims.push(new Zemu(m.path))
-      else if (ONE_APP_PER_PARTICIPANT) for (let i = 0; i < participants; i++) globalSims.push(new Zemu(m.path))
-
-      for (let i = 0; i < globalSims.length; i++)
-        await globalSims[i].start({
-          ...defaultOptions,
-          model: m.name,
-          startText: startTextFn(m.name),
-          approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-          approveAction: ButtonKind.ApproveTapButton,
-        })
-
-      try {
-        for (let i = 0; i < participants; i++) {
-          await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            let result = app.dkgRestoreKeys(encrypted[i])
-
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-sign-${index}-restore-keys`)
-
-            await result
-          })
-        }
-
-        let viewKey = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result: any = await app.dkgRetrieveKeys(IronfishKeys.ViewKey)
-
-          return {
-            viewKey: result.viewKey.toString('hex'),
-            ivk: result.ivk.toString('hex'),
-            ovk: result.ovk.toString('hex'),
-          }
-        })
-
-        let proofKey = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result: any = await app.dkgRetrieveKeys(IronfishKeys.ProofGenerationKey)
-
-          return { ak: result.ak.toString('hex'), nsk: result.nsk.toString('hex') }
-        })
-
-        let pubkey = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result: any = await app.dkgRetrieveKeys(IronfishKeys.PublicAddress)
-
-          return result.publicAddress.toString('hex')
-        })
-
-        let publicPackage = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result = await app.dkgGetPublicPackage()
-
-          return result.publicPackage.toString('hex')
-        })
-
-        for (let i = 0; i < participants; i++) {
-          const identity = await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            return await app.dkgRetrieveKeys(IronfishKeys.DkgIdentity)
-          })
-
-          if (!identity.identity) throw new Error('no identity found')
-
-          identities.push(identity.identity.toString('hex'))
-        }
-
-        const unsignedTxRaw = buildTx(pubkey, viewKey, proofKey)
-        const unsignedTx = new UnsignedTransaction(unsignedTxRaw)
-
-        const serialized = unsignedTx.serialize()
-
-        for (let i = 0; i < participants; i++) {
-          await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            // Change the approve button type to hold, as we are signing a tx now.
-            sim.startOptions.approveAction = ButtonKind.ApproveHoldButton
-            const resultReq = app.reviewTransaction(serialized.toString('hex'))
-
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-sign-${index}-review-transaction`)
-
-            const result = await resultReq
-            expect(result.hash.length).toBeTruthy()
-            expect(result.hash.toString('hex')).toBe(unsignedTx.hash().toString('hex'))
-
-            return result
-          })
-        }
-
-        for (let i = 0; i < participants; i++) {
-          const result = await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            let result = await app.dkgGetCommitments(unsignedTx.hash().toString('hex'))
-
-            expect(result.commitments.length).toBeTruthy()
-
-            return result
-          })
-
-          commitments.push(
-            multisig.SigningCommitment.fromRaw(identities[i], result.commitments, unsignedTx.hash(), identities)
-              .serialize()
-              .toString('hex'),
-          )
-        }
-
-        const signingPackageHex = unsignedTx.signingPackage(commitments)
-        const signingPackage = new multisig.SigningPackage(Buffer.from(signingPackageHex, 'hex'))
-
-        for (let i = 0; i < participants; i++) {
-          const result = await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            let result = await app.dkgSign(
-              unsignedTx.publicKeyRandomness(),
-              signingPackage.frostSigningPackage().toString('hex'),
-              unsignedTx.hash().toString('hex'),
-            )
-
-            expect(result.signature.length).toBeTruthy()
-
-            return result
-          })
-
-          signatures.push(
-            multisig.SignatureShare.fromFrost(result.signature, Buffer.from(identities[i], 'hex')).serialize().toString('hex'),
-          )
-        }
-
-        // Attempt to sign again. It should fail as the tx hash is cleaned
-        for (let i = 0; i < participants; i++) {
-          await expect(
-            runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-              await app.dkgSign(
-                unsignedTx.publicKeyRandomness(),
-                signingPackage.frostSigningPackage().toString('hex'),
-                unsignedTx.hash().toString('hex'),
-              )
-            }),
-          ).rejects.toThrow()
-        }
-
-        let signedTxRaw = aggregateSignatureShares(publicPackage, signingPackageHex, signatures)
-        expect(verifyTransactions([signedTxRaw])).toBeTruthy()
-
-        const signedTx = new Transaction(signedTxRaw)
-        expect(signedTx.spends.length).toBe(1)
-        expect(signedTx.mints.length).toBe(1)
-        expect(signedTx.burns.length).toBe(0)
-      } finally {
-        for (let i = 0; i < globalSims.length; i++) await globalSims[i].close()
-      }
-    })
-  })
-
-  describe.each(restoreKeysTestCases)(`${m.name} - attempt to sign after sending wrong command`, ({ index, encrypted }) => {
-    test(index + '', async () => {
-      const participants = encrypted.length
-      const globalSims: Zemu[] = []
-
-      let identities: any[] = []
-
-      if (ONE_GLOBAL_APP) globalSims.push(new Zemu(m.path))
-      else if (ONE_APP_PER_PARTICIPANT) for (let i = 0; i < participants; i++) globalSims.push(new Zemu(m.path))
-
-      for (let i = 0; i < globalSims.length; i++)
-        await globalSims[i].start({
-          ...defaultOptions,
-          model: m.name,
-          startText: startTextFn(m.name),
-          approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-          approveAction: ButtonKind.ApproveTapButton,
-        })
-
-      try {
-        const reqs = []
-        for (let i = 0; i < participants; i++) {
-          await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            let result = app.dkgRestoreKeys(encrypted[i])
-
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-sign-${index}-restore-keys`)
-
-            await result
-          })
-        }
-
-        let viewKey = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result: any = await app.dkgRetrieveKeys(IronfishKeys.ViewKey)
-
-          return {
-            viewKey: result.viewKey.toString('hex'),
-            ivk: result.ivk.toString('hex'),
-            ovk: result.ovk.toString('hex'),
-          }
-        })
-
-        let proofKey = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result: any = await app.dkgRetrieveKeys(IronfishKeys.ProofGenerationKey)
-
-          return { ak: result.ak.toString('hex'), nsk: result.nsk.toString('hex') }
-        })
-
-        let pubkey = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result: any = await app.dkgRetrieveKeys(IronfishKeys.PublicAddress)
-
-          return result.publicAddress.toString('hex')
-        })
-
-        let publicPackages = await runMethod(m, globalSims, 0, async (sim: Zemu, app: IronfishApp) => {
-          let result = await app.dkgGetPublicPackage()
-
-          return result.publicPackage
-        })
-
-        for (let i = 0; i < participants; i++) {
-          const identity = await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            return await app.dkgRetrieveKeys(IronfishKeys.DkgIdentity)
-          })
-
-          if (!identity.identity) throw new Error('no identity found')
-
-          identities.push(identity.identity.toString('hex'))
-        }
-
-        const unsignedTxRaw = buildTx(pubkey, viewKey, proofKey)
-        const unsignedTx = new UnsignedTransaction(unsignedTxRaw)
-
-        const serialized = unsignedTx.serialize()
-
-        for (let i = 0; i < participants; i++) {
-          await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            // Change the approve button type to hold, as we are signing a tx now.
-            sim.startOptions.approveAction = ButtonKind.ApproveHoldButton
-            const resultReq = app.reviewTransaction(serialized.toString('hex'))
-
-            await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-            await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-sign-${index}-review-transaction`)
-
-            const result = await resultReq
-            expect(result.hash.length).toBeTruthy()
-            expect(result.hash.toString('hex')).toBe(unsignedTx.hash().toString('hex'))
-
-            return result
-          })
-        }
-
-        // Send wrong command in the middle of signing process (review + commitments + sign)
-        for (let i = 0; i < participants; i++) {
-          const identity = await runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-            return await app.dkgRetrieveKeys(IronfishKeys.DkgIdentity)
-          })
-
-          if (!identity.identity) throw new Error('no identity found')
-        }
-
-        // Attempt to get commitments
-        for (let i = 0; i < participants; i++) {
-          await expect(
-            runMethod(m, globalSims, i, async (sim: Zemu, app: IronfishApp) => {
-              let result = await app.dkgGetCommitments(unsignedTx.hash().toString('hex'))
-
-              expect(result.commitments.length).toBeTruthy()
-
-              return result
-            }),
-          ).rejects.toThrow()
-        }
-      } finally {
-        for (let i = 0; i < globalSims.length; i++) await globalSims[i].close()
-      }
-    })
-  })
-
-  test.concurrent(`${m.name} - attempt to retrieve viewKeys when no keys are present`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      await expect(app.dkgRetrieveKeys(IronfishKeys.ViewKey)).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-
-  // TODO implement a way to send the command, and but no get the response
-  /*
-  test.concurrent(`${m.name} - attempt to retrieve result after another command`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      let respReq = app.dkgBackupKeys()
-
-      await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot())
-      await sim.compareSnapshotsAndApprove('.', `${m.prefix.toLowerCase()}-dkg-d`)
-
-      const resp = await respReq
-    } finally {
-      await sim.close()
-    }
-  })
-  */
-
-  test.concurrent(`${m.name} - attempt to retrieve proof keys when no keys are present`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      await expect(app.dkgRetrieveKeys(IronfishKeys.ProofGenerationKey)).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-
-  test.concurrent(`${m.name} - attempt to retrieve public address when no keys are present`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      await expect(app.dkgRetrieveKeys(IronfishKeys.PublicAddress)).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-
-  test.concurrent(`${m.name} - attempt to retrieve public package when no keys are present`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      await expect(app.dkgGetPublicPackage()).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-
-  test.concurrent(`${m.name} - attempt to backup keys when no keys are present`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      await expect(app.dkgBackupKeys()).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-
-  test.concurrent(`${m.name} - attempt to run round1 with 5 participants`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({
-        ...defaultOptions,
-        model: m.name,
-        startText: startTextFn(m.name),
-        approveKeyword: isTouchDevice(m.name) ? 'Approve' : '',
-        approveAction: ButtonKind.ApproveTapButton,
-      })
-      const app = new IronfishApp(sim.getTransport(), true)
-
-      await expect(
-        app.dkgRound1(
-          0,
-          identities.map(({ v }) => v),
-          3,
-        ),
-      ).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-
-  // TODO complete me
-  /*
-  test.concurrent(`${m.name} - attempt to run round3 when no round1 was executed`, async () => {
-    const sim = new Zemu(m.path)
-    try {
-      await sim.start({ ...defaultOptions, model: m.name, startText: startTextFn(m.name) })
-      const app = new IronfishApp(sim.getTransport(), true)
-      let resp: any = await app.dkgRound3()
-
-      await expect(app.dkgRound3()).rejects.toThrow()
-    } finally {
-      await sim.close()
-    }
-  })
-  */
 })
